backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
cffi==1.17.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
django[argon2]==5.1; python_version >= '3.10'
django-allauth[socialaccount]==64.0.0; python_version >= '3.8'
django-allauth-ui==1.4.0; python_version >= '3.8' and python_full_version < '4.0.0'
django-ninja==1.3.0; python_version >= '3.7'
django-widget-tweaks==1.5.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
psycopg[binary]==3.2.1; python_version >= '3.8'
psycopg-binary==3.2.1
pycparser==2.22; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
slippers==0.6.2; python_full_version >= '3.8.0'
sqlparse==0.5.1; python_version >= '3.8'
typeguard==2.13.3; python_full_version >= '3.5.3'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.30.5; python_version >= '3.8'
